Move map_x functions to conv.rs
